+++ activesupport/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Raise `ArgumentError` in `ActiveSupport::TimeZone#parse` when an invalid string is provided, aligning with documentation.
+
+    *Koji Toga*
+
 *   Add `ActiveSupport::CombinedConfiguration` to offer interchangeable access to configuration provided by
     either ENV or encrypted credentials. Used by Rails to first look at ENV, then look in encrypted credentials,
     but can be configured separately with any number of API-compatible backends in a first-look order.